bsunanda:Run2-alca62 Modify TDC mask in HcalTPParameters from 32 bit to 64 bit uint

+++ CondFormats/HcalObjects/interface/HcalTPParameters.h
@@ -14,17 +14,17 @@ public:
   ~HcalTPParameters();
 
   // Load a new entry
-  void         loadObject(int version, int adcCut, uint32_t tdcMask,
-			  uint32_t tbits, int auxi1,  int auxi2);
+  void         loadObject(int version, int adcCut, uint64_t tdcMask,
+			  uint32_t tbits, int auxi1, int auxi2);
 
   /// get FineGrain Algorithm Version for HBHE
   int          getFGVersionHBHE()          const {return version_;}
   /// get ADC threshold fof TDC mask of HF
   int          getADCThresholdHF()         const {return adcCut_;}
   /// get TDC mask for HF
-  uint32_t getTDCMaskHF()                  const {return tdcMask_;}
+  uint64_t     getTDCMaskHF()              const {return tdcMask_;}
   /// get Self Trigger bits
-  uint32_t getHFTriggerInfo()              const {return tbits_;}
+  uint32_t     getHFTriggerInfo()          const {return tbits_;}
   /// get Axiliary words
   int          getAuxi1()                  const {return auxi1_;}
   int          getAuxi2()                  const {return auxi2_;}
@@ -33,7 +33,7 @@ private:
 
   int          version_;
   int          adcCut_;
-  uint32_t     tdcMask_;
+  uint64_t     tdcMask_;
   uint32_t     tbits_;
   int          auxi1_;
   int          auxi2_;
